requirements.txt
matplotlib==3.6.1
pandas==1.5.1
cvxpy==1.2.2
xlsxwriter==3.0.3
openpyxl==3.0.10
PyYAML